Added screen buffer, altered interfaces and tests accordingly

+++ src/test/TestPoint.java
@@ -7,7 +7,7 @@ public class TestPoint implements PointInterface {
 	public double x;
 	public double y;
 	
-	public TestPoint(int x, int y) {
+	public TestPoint(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
